refactor(data_classes): rename package from trigger
