+++ aws_lambda_powertools/utilities/data_classes/api_gateway_proxy_event.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, List, Optional
 
-from aws_lambda_powertools.utilities.trigger.common import BaseProxyEvent, DictWrapper
+from aws_lambda_powertools.utilities.data_classes.common import BaseProxyEvent, DictWrapper
 
 
 class APIGatewayEventIdentity(DictWrapper):
