+++ aws_lambda_powertools/utilities/data_classes/dynamo_db_stream_event.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Dict, Iterator, List, Optional
 
-from aws_lambda_powertools.utilities.trigger.common import DictWrapper
+from aws_lambda_powertools.utilities.data_classes.common import DictWrapper
 
 
 class AttributeValue(DictWrapper):
