+++ aws_lambda_powertools/utilities/data_classes/s3_event.py
@@ -1,7 +1,7 @@
 from typing import Dict, Iterator, Optional
 from urllib.parse import unquote_plus
 
-from aws_lambda_powertools.utilities.trigger.common import DictWrapper
+from aws_lambda_powertools.utilities.data_classes.common import DictWrapper
 
 
 class S3Identity(DictWrapper):
