+++ aws_lambda_powertools/utilities/data_classes/sqs_event.py
@@ -1,6 +1,6 @@
 from typing import Dict, Iterator, Optional
 
-from aws_lambda_powertools.utilities.trigger.common import DictWrapper
+from aws_lambda_powertools.utilities.data_classes.common import DictWrapper
 
 
 class SQSRecordAttributes(DictWrapper):
